src/backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Database dependencies
sqlalchemy>=2.0.0
alembic>=1.10.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# AI/ML dependencies
google-generativeai>=0.3.0

# Configuration and utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.5

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
httpx>=0.24.0

# Development dependencies
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

